Update d2d2-core to 0.1.5-SNAPSHOT

+++ d2d2-world/src/main/java/com/ancevt/d2d2world/gameobject/IRepeatable.java
@@ -20,19 +20,19 @@ package com.ancevt.d2d2world.gameobject;
 import com.ancevt.d2d2world.data.Property;
 
 public interface IRepeatable {
-	void setRepeat(int repeatX, int repeatY);
+	void setRepeat(float repeatX, float repeatY);
 
 	@Property
-	void setRepeatX(int repeatX);
+	void setRepeatX(float repeatX);
 
 	@Property
-	void setRepeatY(int repeatY);
+	void setRepeatY(float repeatY);
 
 	@Property
-	int getRepeatX();
+	float getRepeatX();
 
 	@Property
-	int getRepeatY();
+	float getRepeatY();
 
 	float getOriginalWidth();
 	float getOriginalHeight();
